fix(numeric): explicit cast in convert_to_decimal bit-test (-Wsign-conversion) (#1270) (#1271)
number/support/decimal.hpp: `if (v & mask)` in convert_to_decimal(long long v)
implicitly promoted the signed `long long v` to `unsigned long long` (mask is
uint64_t), tripping -Wsign-conversion. `v` is non-negative at that point -- the
function transforms negatives to positive magnitude up front (v *= -1) -- so the
bit-test is correct; make the intent explicit with static_cast<uint64_t>(v),
matching the existing decimal::setvalue idiom (`if (absValue & mask)`).

Value-preserving. decimal.hpp is now clean under -Wconversion / -Wsign-conversion
/ -Wignored-qualifiers on gcc + clang; conversions_to_decimal and bt_decimal PASS
on both.

Resolves #1270

+++ include/sw/universal/number/support/decimal.hpp
@@ -383,7 +383,7 @@ inline void convert_to_decimal(long long v, decimal& d) {
 	decimal base; // can't use base(1) semantics here as it would cause an infinite loop
 	base.setdigit(1);
 	while (v) { // minimum loop iterations; exits when no bits left
-		if (v & mask) {
+		if (static_cast<uint64_t>(v) & mask) {   // v is non-negative here (magnitude); cast avoids -Wsign-conversion (#1270)
 			add(d, base);
 		}
 		add(base, base);
